front_and_agent/requirements.txt
streamlit
requests
python-dotenv
fastapi
uvicorn[standard]
httpx
langid
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
langchain-text-splitters
transformers
torch
sentence-transformers
faiss-cpu
pydantic
